# PERF: int nansum contiguous case